Separate executor and scala future tests
