+++ dd-java-agent/instrumentation/java-concurrent/src/test/groovy/ExecutorInstrumentationTest.groovy
@@ -118,73 +118,4 @@ class ExecutorInstrumentationTest extends AgentTestRunner {
     new ThreadPoolExecutor(1, 1, 1000, TimeUnit.NANOSECONDS, new ArrayBlockingQueue<Runnable>(1))   | _
     new ScheduledThreadPoolExecutor(1)                                                              | _
   }
-
-  def "scala futures and callbacks"() {
-    setup:
-    ScalaConcurrentTests scalaTest = new ScalaConcurrentTests()
-    int expectedNumberOfSpans = scalaTest.traceWithFutureAndCallbacks()
-    TEST_WRITER.waitForTraces(1)
-    List<DDSpan> trace = TEST_WRITER.get(0)
-
-    expect:
-    trace.size() == expectedNumberOfSpans
-    trace[0].operationName == "ScalaConcurrentTests.traceWithFutureAndCallbacks"
-    findSpan(trace, "goodFuture").context().getParentId() == trace[0].context().getSpanId()
-    findSpan(trace, "badFuture").context().getParentId() == trace[0].context().getSpanId()
-    findSpan(trace, "successCallback").context().getParentId() == trace[0].context().getSpanId()
-    findSpan(trace, "failureCallback").context().getParentId() == trace[0].context().getSpanId()
-  }
-
-  def "scala propagates across futures with no traces"() {
-    setup:
-    ScalaConcurrentTests scalaTest = new ScalaConcurrentTests()
-    int expectedNumberOfSpans = scalaTest.tracedAcrossThreadsWithNoTrace()
-    TEST_WRITER.waitForTraces(1)
-    List<DDSpan> trace = TEST_WRITER.get(0)
-
-    expect:
-    trace.size() == expectedNumberOfSpans
-    trace[0].operationName == "ScalaConcurrentTests.tracedAcrossThreadsWithNoTrace"
-    findSpan(trace, "callback").context().getParentId() == trace[0].context().getSpanId()
-  }
-
-  def "scala either promise completion"() {
-    setup:
-    ScalaConcurrentTests scalaTest = new ScalaConcurrentTests()
-    int expectedNumberOfSpans = scalaTest.traceWithPromises()
-    TEST_WRITER.waitForTraces(1)
-    List<DDSpan> trace = TEST_WRITER.get(0)
-
-    expect:
-    TEST_WRITER.size() == 1
-    trace.size() == expectedNumberOfSpans
-    trace[0].operationName == "ScalaConcurrentTests.traceWithPromises"
-    findSpan(trace, "keptPromise").context().getParentId() == trace[0].context().getSpanId()
-    findSpan(trace, "keptPromise2").context().getParentId() == trace[0].context().getSpanId()
-    findSpan(trace, "brokenPromise").context().getParentId() == trace[0].context().getSpanId()
-  }
-
-  def "scala first completed future"() {
-    setup:
-    ScalaConcurrentTests scalaTest = new ScalaConcurrentTests()
-    int expectedNumberOfSpans = scalaTest.tracedWithFutureFirstCompletions()
-    TEST_WRITER.waitForTraces(1)
-    List<DDSpan> trace = TEST_WRITER.get(0)
-
-    expect:
-    TEST_WRITER.size() == 1
-    trace.size() == expectedNumberOfSpans
-    findSpan(trace, "timeout1").context().getParentId() == trace[0].context().getSpanId()
-    findSpan(trace, "timeout2").context().getParentId() == trace[0].context().getSpanId()
-    findSpan(trace, "timeout3").context().getParentId() == trace[0].context().getSpanId()
-  }
-
-  private DDSpan findSpan(List<DDSpan> trace, String opName) {
-    for (DDSpan span : trace) {
-      if (span.getOperationName() == opName) {
-        return span
-      }
-    }
-    return null
-  }
 }
